rearrange some functions
