+++ py_sst_to_cam/sst_to_cam.py
@@ -11,54 +11,8 @@ module_path = os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'py_
 if module_path not in sys.path:
     sys.path.append(module_path)
 import pyfuncs
+import horizremap
 
-def linint2(xi, yi, fi, fiCyclicX, xo, yo):
-    """
-    Bilinear interpolation from one rectilinear grid to another.
-
-    Parameters:
-    - xi: 1D or 2D array of X coordinates of the input data.
-    - yi: 1D or 2D array of Y coordinates of the input data.
-    - fi: 2D or 3D array of input data to be interpolated.
-    - fiCyclicX: Boolean indicating if the X dimension is cyclic.
-    - xo: 1D array of X coordinates of the output data.
-    - yo: 1D array of Y coordinates of the output data.
-
-    Returns:
-    - Interpolated array of the same dimensions as fi, except the last two dimensions.
-    """
-
-    # Handle cyclic conditions
-    if fiCyclicX:
-        # Extend the fi array and xi array for cyclic boundary
-        xi_extended = np.concatenate((xi, xi[:1] + 360.0), axis=-1)
-        fi_extended = np.concatenate((fi, fi[..., :1]), axis=-1)
-        xi = xi_extended
-        fi = fi_extended
-
-    # Create the interpolator object
-    interpolator = RegularGridInterpolator((yi, xi), fi, method='linear', bounds_error=False, fill_value=np.nan)
-
-    # Create a meshgrid for target coordinates
-    xo_mesh, yo_mesh = np.meshgrid(xo, yo, indexing='xy')
-
-    # Prepare the points for interpolation
-    target_points = np.vstack((yo_mesh.ravel(), xo_mesh.ravel())).T
-
-    # Perform the interpolation
-    interpolated_data = interpolator(target_points)
-
-    # Reshape the interpolated data to match the target grid shape
-    interpolated_data = interpolated_data.reshape(yo_mesh.shape)
-
-    return interpolated_data
-
-def parse_initdate(initdate):
-    yyyy = int(initdate[:4])
-    mm = int(initdate[4:6])
-    dd = int(initdate[6:8])
-    hh = int(initdate[8:])
-    return yyyy, mm, dd, hh
 
 def process_sst_ice(initdate, predict_docn, inputres, datasource, sstDataFile, iceDataFile, SST_write_file):
     # Constants
@@ -74,7 +28,7 @@ def process_sst_ice(initdate, predict_docn, inputres, datasource, sstDataFile, i
     do_anom = predict_docn == 1
     logging.info(f"do_anom (i.e., predict_docn) set to: {do_anom}")
 
-    yyyy, mm, dd, hh = parse_initdate(initdate)
+    yyyy, mm, dd, hh = pyfuncs.parse_YYYYMMDDHH(initdate)
 
     logging.info(f"running date {initdate}")
 
@@ -118,15 +72,15 @@ def process_sst_ice(initdate, predict_docn, inputres, datasource, sstDataFile, i
     fvlon = in_ds['xc'].values
 
     if datasource == "GDAS":
-        sst_fv = linint2(sstlon, sstlat[::-1], sst_gfs[::-1], True, fvlon, fvlat)
-        ice_fv = linint2(sstlon, sstlat[::-1], ice_gfs[::-1], True, fvlon, fvlat)
+        sst_fv = horizremap.linint2(sstlon, sstlat[::-1], sst_gfs[::-1], True, fvlon, fvlat)
+        ice_fv = horizremap.linint2(sstlon, sstlat[::-1], ice_gfs[::-1], True, fvlon, fvlat)
     elif datasource == "NOAAOI":
-        sst_fv = linint2(sstlon, sstlat, sst_gfs, True, fvlon, fvlat)
-        ice_fv = linint2(sstlon, sstlat, ice_gfs, True, fvlon, fvlat)
+        sst_fv = horizremap.linint2(sstlon, sstlat, sst_gfs, True, fvlon, fvlat)
+        ice_fv = horizremap.linint2(sstlon, sstlat, ice_gfs, True, fvlon, fvlat)
 
     # Handling missing values
-    sst_fv = np.where(np.isnan(sst_fv), TTHRESH, sst_fv)
-    sst_fv = np.where(sst_fv > 500, TTHRESH, sst_fv)
+    sst_fv = np.where(np.isnan(sst_fv), (TTHRESH-KtoC), sst_fv)
+    sst_fv = np.where(sst_fv > 500, (TTHRESH-KtoC), sst_fv)
 
     # SST to double precision
     sst_fv_dbl = sst_fv.astype(np.float64)
@@ -196,45 +150,12 @@ def process_sst_ice(initdate, predict_docn, inputres, datasource, sstDataFile, i
     out_ds.to_netcdf(SST_write_file)
     logging.info(f"Data written to {SST_write_file}")
 
-def parse_args_sst():
-    parser = argparse.ArgumentParser(description="Process SST and Ice data for CESM F compsets.")
-
-    # Required arguments
-    parser.add_argument('--initdate', type=str, required=True,
-                        help='Initialization date in format YYYYMMDDHH')
-    parser.add_argument('--predict_docn', type=int, choices=[0, 1], required=True,
-                        help='Flag for predict_docn: 0 (false) or 1 (true)')
-    parser.add_argument('--inputres', type=str, required=True,
-                        help='Input resolution (e.g., "180x360")')
-    parser.add_argument('--datasource', type=str, required=True,
-                        help='Data source (e.g., "GDAS", "NOAAOI")')
-    parser.add_argument('--sstDataFile', type=str, required=True,
-                        help='Full path to the SST data file')
-    parser.add_argument('--iceDataFile', type=str, required=True,
-                        help='Full path to the Ice data file')
-    parser.add_argument('--SST_write_file', type=str, required=True,
-                        help='Full path to the output SST file')
-
-    # Optional arguments with defaults
-    parser.add_argument('--smooth_ice', action='store_true', default=False,
-                        help='If set, smooth the ice field (default: False)')
-    parser.add_argument('--smooth_iter', type=int, default=3,
-                        help='Number of iterations for smoothing the ice field (default: 3)')
-    parser.add_argument('--TTHRESH', type=float, default=271.9,
-                        help='Initial threshold for ice vs. open ocean (default: 271.9)')
-    parser.add_argument('--KtoC', type=float, default=273.15,
-                        help='Kelvin to Celsius conversion constant (default: 273.15)')
-    parser.add_argument('-v', '--verbose', action='store_true',
-                        help='Increase output verbosity for debugging')
-
-    return parser.parse_args()
 
 def main():
-    args = parse_args_sst()
+    args = pyfuncs.parse_args_sst()
 
     pyfuncs.configure_logging(args.verbose)
 
-    # Pass the parsed arguments to your processing function
     process_sst_ice(
         args.initdate,
         args.predict_docn,
